+++ test/distributed/test_dist_neighbor_sampler.py
@@ -76,7 +76,6 @@ def create_data(rank: int, world_size: int, time_attr: Optional[str] = None):
 def create_hetero_data(
     tmp_path: str,
     rank: int,
-    time_attr: Optional[str] = None,
 ):
     graph_store = LocalGraphStore.from_partition(tmp_path, pid=rank)
     feature_store = LocalFeatureStore.from_partition(tmp_path, pid=rank)
@@ -92,28 +91,6 @@ def create_hetero_data(
     graph_store.node_pb = feature_store.node_feat_pb = node_pb
     graph_store.edge_pb = feature_store.edge_feat_pb = edge_pb
     graph_store.meta = feature_store.meta = meta
-
-    if time_attr == 'time':  # Create node-level time data:
-        feature_store.put_tensor(
-            tensor=torch.full((len(node_pb['v0']), ), 1, dtype=torch.int64),
-            group_name='v0',
-            attr_name=time_attr,
-        )
-        feature_store.put_tensor(
-            tensor=torch.full((len(node_pb['v1']), ), 2, dtype=torch.int64),
-            group_name='v1',
-            attr_name=time_attr,
-        )
-    elif time_attr == 'edge_time':  # Create edge-level time data:
-        i = 0
-        for attr, edge_index in graph_store._edge_index.items():
-            time = torch.full((edge_index.size(1), ), i, dtype=torch.int64)
-            feature_store.put_tensor(
-                tensor=time,
-                group_name=attr[0],
-                attr_name=time_attr,
-            )
-            i += 1
 
     return feature_store, graph_store
 
@@ -350,7 +327,7 @@ def dist_neighbor_sampler_temporal_hetero(
     temporal_strategy: str = 'uniform',
     time_attr: str = 'time',
 ):
-    dist_data = create_hetero_data(tmp_path, rank, time_attr)
+    dist_data = create_hetero_data(tmp_path, rank)
 
     current_ctx = DistContext(
         rank=rank,
@@ -575,15 +552,13 @@ def test_dist_neighbor_sampler_temporal_hetero(
         edge_dim=2,
     )[0]
 
-    partitioner = Partitioner(data, world_size, tmp_path)
-    partitioner.generate_partition()
-
-    # The partition generation script does not currently support temporal data.
-    # Therefore, it needs to be added after generating partitions.
     data['v0'].time = torch.full((data.num_nodes_dict['v0'], ), 1,
                                  dtype=torch.int64)
     data['v1'].time = torch.full((data.num_nodes_dict['v1'], ), 2,
                                  dtype=torch.int64)
+
+    partitioner = Partitioner(data, world_size, tmp_path)
+    partitioner.generate_partition()
 
     w0 = mp_context.Process(
         target=dist_neighbor_sampler_temporal_hetero,
@@ -629,14 +604,12 @@ def test_dist_neighbor_sampler_edge_level_temporal_hetero(
         edge_dim=2,
     )[0]
 
-    partitioner = Partitioner(data, world_size, tmp_path)
-    partitioner.generate_partition()
-
-    # The partition generation script does not currently support temporal data.
-    # Therefore, it needs to be added after generating partitions.
     for i, edge_type in enumerate(data.edge_types):
         data[edge_type].edge_time = torch.full(
             (data[edge_type].edge_index.size(1), ), i, dtype=torch.int64)
+
+    partitioner = Partitioner(data, world_size, tmp_path)
+    partitioner.generate_partition()
 
     w0 = mp_context.Process(
         target=dist_neighbor_sampler_temporal_hetero,