+++ test/nn/conv/test_cluster_gcn_conv.py
@@ -23,7 +23,7 @@ def test_cluster_gcn_conv():
         assert torch.allclose(conv(x, adj2.t()), out, atol=1e-5)
 
     if is_full_test():
-        jit = torch.jit.script(conv.jittable())
+        jit = torch.jit.script(conv)
         assert torch.allclose(jit(x, edge_index), out)
 
         if torch_geometric.typing.WITH_TORCH_SPARSE: