+++ test/nn/models/test_rect.py
@@ -32,7 +32,7 @@ def test_rect():
     assert labels_out.size() == (int(mask.sum()), 8)
 
     if is_full_test():
-        jit = torch.jit.script(model.jittable())
+        jit = torch.jit.script(model)
         assert torch.allclose(jit(x, edge_index), out, atol=1e-6)
         assert torch.allclose(embed_out, jit.embed(x, edge_index), atol=1e-6)
         assert torch.allclose(labels_out, jit.get_semantic_labels(x, y, mask))