+++ torch_geometric/nn/conv/gated_graph_conv.py
@@ -4,7 +4,7 @@ from torch.nn import Parameter as Param
 
 from torch_geometric.nn.conv import MessagePassing
 from torch_geometric.nn.inits import uniform
-from torch_geometric.typing import Adj, OptTensor, SparseTensor
+from torch_geometric.typing import Adj, OptTensor
 from torch_geometric.utils import spmm
 
 
@@ -84,7 +84,7 @@ class GatedGraphConv(MessagePassing):
     def message(self, x_j: Tensor, edge_weight: OptTensor):
         return x_j if edge_weight is None else edge_weight.view(-1, 1) * x_j
 
-    def message_and_aggregate(self, adj_t: SparseTensor, x: Tensor) -> Tensor:
+    def message_and_aggregate(self, adj_t: Adj, x: Tensor) -> Tensor:
         return spmm(adj_t, x, reduce=self.aggr)
 
     def __repr__(self) -> str: